Refactor / code cleanup from auth and adding likes/bookmark routes
