+++ components/Authentication.js
@@ -2,11 +2,9 @@ import React, { Component } from 'react';
 import { StyleSheet, Text, View, AsyncStorage, TextInput, Platform, Button, Alert} from 'react-native';
 import { connect } from 'react-redux';
 import { actionCreators } from '../actions';
-const options = {};
-
+import { _userSignup(), _userLogin } from '../helpers':
 const mapStateToProps = (state) => ({
-  view: state.main.view,
-  token: state.main.token
+  view: state.main.view
 })
 
 class Authentication extends Component {
@@ -24,87 +22,9 @@ class Authentication extends Component {
   async _onValueChange(item, selectedValue) {
     try {
       await AsyncStorage.setItem(item, selectedValue);
-      this.props.dispatch(actionCreators.addtoken(selectedValue))
     } catch (error) {
       console.log('AsyncStorage error: ' + error.message);
     }
-  }
-
-  _userSignup() {
-    if(this.state.password === '' || this.state.username === '' || this.state.email === '') {
-      Alert.alert('Missing fields');
-      return;
-    }
-    if(this.state.password.length < 6) {
-      Alert.alert('Password must be at least 6 characters long.');
-      return;
-    }
-    if(!this.state.email.includes('@')) {
-      Alert.alert('Please enter a valid email.');
-      return;
-    }
-    var value = {username: this.state.username, password: this.state.password, email: this.state.email, avatarUrl: this.state.avatarUrl};
-      fetch("http:localhost:3000/api/users/createUser", {
-        method: "POST",
-        headers: {
-          'Accept': 'application/json',
-          'Content-Type': 'application/json'
-        },
-        body: JSON.stringify(value)
-      })
-      .then((response) => {
-        if (response.status !== 201) {
-          response.json()
-          .then((responseData) => Alert.alert(responseData.message))
-          .catch(console.warn);
-        } else {
-          return response.json();
-        }
-      })
-      .then((responseData) => {
-        if (responseData) {
-          this.props.dispatch(actionCreators.changeView('Inbox'))
-          return this._onValueChange('id_token', responseData.id_token)
-        }
-      })
-      .catch((error) => {
-        console.warn('Error: ', error);
-      })
-      .done();
-  }
-
-  _userLogin() {
-    if(this.state.password === '' || this.state.username === '') {
-      Alert.alert('Missing fields');
-      return;
-    }
-    var value = {username: this.state.username, password: this.state.password};
-      fetch("http://localhost:3000/api/users/login", {
-        method: "POST",
-        headers: {
-          'Accept': 'application/json',
-          'Content-Type': 'application/json'
-        },
-        body: JSON.stringify(value)
-      })
-       .then((response) => {
-        if (response.status !== 201) {
-          response.json()
-          .then((responseData) => Alert.alert(responseData.message))
-        } else {
-          return response.json();
-        }
-      })
-      .then((responseData) => {
-        if (responseData) {
-          this.props.dispatch(actionCreators.changeView('Inbox'))
-          return this._onValueChange('id_token', responseData.id_token)
-        }
-      })
-      .catch((error) => {
-        console.warn('Error: ', error);
-      })
-      .done();
   }
 
   render() {
